requirements.txt
mcp[sse]
uvicorn
requests
web3
supabase
python-dotenv